requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.9.3
cachetools==5.5.2
certifi==2021.5.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.4
django-cors-headers==4.7.0
django-ipware==7.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
geoip2==5.0.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
maxminddb==2.6.3
multidict==6.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-ipware==3.0.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.2.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.2.1
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
yarl==1.18.3
mysqlclient